+++ packages/vm/src/lib.rs
@@ -2,6 +2,8 @@ pub mod cache;
 pub mod calls;
 pub mod compile;
 pub mod error;
+mod imports;
+mod store;
 pub mod vm;
 
 pub use calls::run;
